Add migration change value Integer on EmergencyContacto to BigInteger

+++ src/emergency_contacts/models.py
@@ -15,7 +15,7 @@ class EmergencyContact(TimeStampAbstract):
         'Nombre completo',
         max_length=128
     )
-    mobile = models.PositiveIntegerField(
+    mobile = models.PositiveBigIntegerField(
         'Celular',
     )
     relationship = models.CharField(
